move js target configuration to buildSrc
